requirements.txt
# QuizLM Requirements
# Python 3.12+

# UI Framework
customtkinter==5.2.1

# LLM Clients
anthropic==0.39.0
openai==1.55.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# PDF Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0

